// spacexinterop.client/src/app/shared/interfaces/settings/accessibility-setting-item.interface.ts
import { MatSelectType, SelectionChange } from "../../types/ui.types";

export interface IAccessibilitySettingItem {
    selfIdentifier: string;

    state?: boolean;

    title?: string;
    matIcons?: string;

    ariaLabel?: string;

    type?: MatSelectType;

    changes?: (event: SelectionChange, value?: any) => any;

    filterStyle?: string;
}